move tests back into cli.rs and move use statement to the top
